feat: latest component structures
