+++ package.json
@@ -52,6 +52,7 @@
     "react-loading-skeleton": "^3.2.0",
     "react-markdown": "^8.0.7",
     "react-router-dom": "^6.9.0",
+    "react-select-country-list": "^2.2.3",
     "react-simplemde-editor": "^5.2.0",
     "react-text-loop": "^2.3.0",
     "react-text-transition": "^3.1.0",
