+++ src/components/Footer.jsx
@@ -15,7 +15,6 @@ export function Footer() {
               <NavLink href="../privacy-policy">Privacy Policy</NavLink>
               <NavLink href="https://status.ctfguide.com/">Status</NavLink>
               <NavLink href="../careers">Careers</NavLink>
-              <NavLink href="../TESTPAGE">TESTPAGE</NavLink>
             </div>
           </nav>
         </div>
