+++ src/pages/explorers.jsx
@@ -16,7 +16,6 @@ import { Logo } from '@/components/Logo'
 const navigation = [
   { name: 'Enterprise', href: 'https://enterprise.ctfguide.com' },
   { name: 'Careers', href: '../careers' },
-  { name: 'TESTPAGE', href: '../TESTPAGE' },
   { name: 'Explorers League', href: 'https://github.com/ctfguide-tech' },
   { name: 'Open Source', href: 'https://github.com/ctfguide-tech' },
 ]
